Try to add mushroom
