+++ CULLinary2/Assets/Game Data/Scriptable Objects/InventoryItems/Mushroom.asset
@@ -16,7 +16,16 @@ MonoBehaviour:
   icon: {fileID: 21300000, guid: b4162b77e2ece47039595440add97ac5, type: 3}
   inventoryItemId: 7
   isConsumable: 0
+  buffIcon: {fileID: 21300000, guid: b4162b77e2ece47039595440add97ac5, type: 3}
   healHpAmount: 0
   increaseMaxHpAmount: 0
   increaseBaseDamageAmount: 0
+  buffDuration: 0
+  isUnlimitedStamina: 0
+  isPlayerInvincible: 0
+  buffBaseDamage: 0
+  buffMoneyBonus: 0
+  evasionBoostAmount: 0
+  critBoostAmount: 0
+  buffTypes: 
   description: 'Warning: Can cause hallucinations if eaten raw'
