+++ CULLinary2/Assets/Models/Mushroom/MushroomDeath.anim
@@ -6474,6 +6474,13 @@ AnimationClip:
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
+  - time: 0.125
+    functionName: destroyHpBarUi
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
   - time: 0.8333333
     functionName: destroyDamageNumbers
     data:
