+++ CULLinary2/Assets/Models/Mushroom/MushroomEnemy.prefab
@@ -1014,8 +1014,8 @@ MonoBehaviour:
   monsterHealth: 100
   distanceTriggered: 10
   stopChase: 14
-  lootDropped: {fileID: 8706522679418664788, guid: bb4dce76efdb34ce682008e7fbc6f1fe, type: 3}
-  additionalSpawningNumbers: 2
+  lootDropped: {fileID: 1302279271560276986, guid: 7923bdb7325a947489c7f86417ee8ec3, type: 3}
+  additionalSpawning: 0
   spawner: {fileID: 0}
   hpBarPrefab: {fileID: 1085209639500972291, guid: 4614e0511cb49b14ebf0c58217d9d222, type: 3}
   damageCounterPrefab: {fileID: 1085209639500972291, guid: 24f41dda89c7bd648bd2796ac669721e, type: 3}
