+++ CULLinary2/Assets/Models/Mushroom/mushroom_texture.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7b9b4e1bf2a8b914fad1ba17613bc3be
+guid: efc90b97955d64340b80eb65ab6bc7c7
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
