+++ CULLinary2/Assets/Models/Potato/TomatoEnemy.prefab
@@ -289,7 +289,7 @@ AudioSource:
   m_Enabled: 1
   serializedVersion: 4
   OutputAudioMixerGroup: {fileID: 0}
-  m_audioClip: {fileID: 8300000, guid: 0a212b4caa7d7be44b784854f3a4eecc, type: 3}
+  m_audioClip: {fileID: 8300000, guid: e0112cdc2521c8a40872086bc6994256, type: 3}
   m_PlayOnAwake: 0
   m_Volume: 0.3
   m_Pitch: 1
