+++ CULLinary2/Assets/Prefabs/Loot/BreadLoot.prefab
@@ -84,6 +84,10 @@ PrefabInstance:
       propertyPath: itemLoot
       value: 
       objectReference: {fileID: 11400000, guid: 749d7c210ea33574a9df743750de6647, type: 2}
+    - target: {fileID: 2087905042338579891, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_Enabled
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 2964862885916105719, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
       propertyPath: m_StaticEditorFlags
       value: 0
