+++ CULLinary2/Assets/Prefabs/Loot/MushroomLoot.prefab
@@ -1,0 +1,137 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &2105791872974939066
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 1093220858637781056, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_Name
+      value: MushroomLoot
+      objectReference: {fileID: 0}
+    - target: {fileID: 1450650355932434205, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: itemLoot
+      value: 
+      objectReference: {fileID: 11400000, guid: 999de4f569e654b98af04bd7b601ea93, type: 2}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6597951860054338853, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+--- !u!4 &7252290103727958480 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 8762889841212541546, guid: d17dfc11e6fc92d43937fe6eff5555c1, type: 3}
+  m_PrefabInstance: {fileID: 2105791872974939066}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &5213081423999159894
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 7252290103727958480}
+    m_Modifications:
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 9
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 9
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 9
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 919132149155446097, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
+      propertyPath: m_Name
+      value: mushroomHalf
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 0bb7e591a6d28854881d4fe4c3fb3c9f, type: 3}
