+++ CULLinary2/Assets/Prefabs/Loot/MushroomLoot.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7923bdb7325a947489c7f86417ee8ec3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
